Внес правки по констрейнам, поправил имя файла ProfileHeaderView, вынес post и posts в отдельный файл
